+++ Common/WebStore.Domain/DTO/SectionDTO.cs
@@ -1,0 +1,10 @@
+﻿namespace WebStore.Domain.DTO
+{
+    public class SectionDTO
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int Order { get; set; }
+        public int? ParentId { get; set; }
+    }
+}
