+++ UI/WebStore/Components/SectionsViewComponent.cs
@@ -25,7 +25,7 @@ namespace WebStore.Components
                    Id = s.Id,
                    Name = s.Name,
                    Order = s.Order,
-                   ProductsCount = s.FromDTO().Products.Count(), // ????
+                   //ProductsCount = s.Products.Count(), // ????
                 })
                 .ToList();
 
@@ -42,7 +42,7 @@ namespace WebStore.Components
                         Name = child_section.Name,
                         Order = child_section.Order,
                         Parent = parent_section,
-                        ProductsCount = child_section.FromDTO().Products.Count(), // ????
+                        //ProductsCount = child_section.Products.Count(), // ????
                     });
                 
                 parent_section.ChildSections.Sort(OrderSortMethod);
